Get `NetworkEndpoint` to use the modern component lifecycle.
